requirements.txt
appscript==1.3.0
blinker==1.9.0
charset-normalizer==3.4.3
click==8.2.1
docx2pdf==0.1.8
fire==0.7.1
Flask==3.1.2
fonttools==4.59.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
pdf2docx==0.5.8
pillow==11.3.0
PyMuPDF==1.26.4
python-docx==1.2.0
reportlab==4.4.3
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.15.0
Werkzeug==3.1.3
